% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.add.r
\name{add_area}
\alias{add_area}
\title{Add areas to a Spatial Polygons Data Frame}
\usage{
add_area(polygons, area_ha = TRUE, area_sqkm = TRUE, area_acres = TRUE)
}
\arguments{
\item{polygons}{An sf polygon object to calculate areas for}

\item{area_ha}{Logical. If \code{TRUE}, areas will be calculated and added in hectares. Default is \code{TRUE}.}

\item{area_sqkm}{Logical. If \code{TRUE}, areas will be calculated and added in square kilometers. Default is \code{TRUE}.}

\item{area_acres}{Logical. If \code{TRUE}, areas will be calculated and added in acres. Default is \code{TRUE}.}
}
\value{
The original sf object with an additional field for each area unit calculated.
}
\description{
This function takes a Spatial Polygons Data Frame and calculates and adds area fields to the data frame. Areas can be calculated either treating the whole SPDF as a unit or for each polygon individually.
}
\examples{
add_area()
}
\keyword{area}
