% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.r
\name{allocate_panels}
\alias{allocate_panels}
\title{Create a design object programmatically to use with \code{spsurvey::grts()}.}
\usage{
allocate_panels(
  spdf,
  stratum_field = "stratum",
  panel_names = c("1", "2", "3", "4", "5"),
  panel_sample_size = 50,
  points_min = 3,
  oversample_proportion = 0.25,
  oversample_min = 3
)
}
\arguments{
\item{spdf}{SpatialPointsDataFrame or data frame containing the possible sampling points being used or a SpatialPolygonsDataFrame of polygons describing the strata boundaries.}

\item{stratum_field}{A character string of the name of the variable in \code{points} containing the stratum identities of the points. Defaults to \code{"stratum"}.}

\item{panel_names}{A character vector of the names to assign to the panels. All values must be unique. This must be the same length and in the same order as \code{panel_sample_size}. Defaults to \code{c("1", "2", "3", "4", "5")}.}

\item{panel_sample_size}{A numeric vector of the number of base points to be drawn in each panel. This must be the same length and in the same order as \code{panel_names} UNLESS all panels have the same number of base points in which case \code{panel_sample_size} optionally may be a single numeric value despite \code{panel_names} containing more than one value. Defaults to \code{c(50)}.}

\item{points_min}{A numeric value of the minimum number of base points to allocate to a stratum within a panel regardless of its relative size. Defaults to \code{3}.}

\item{oversample_proportion}{A numeric value between 0 and 1 representing the minimum relative proportion of oversample points to allocate per stratum per panel using the formula \code{panel_sample_size * min.oversample_proportion}. Defaults to \code{0.25}.}

\item{oversample_min}{A numeric value representing the minimum number of oversample points to allocate per stratum per panel. This is only used if it is greater than \code{panel_sample_size * min.oversample_proportion}. Defaults to \code{3}.}
}
\value{
A named list of named lists conforming to the requirements for the design object for \code{spsurvey::grts()}.
}
\description{
Creates a structured list that meets the requirements for a design object for \code{spsurvey::grts()} distributing points proportionally to strata based on their relative areas/abundances.
}
