% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.shapefile.r
\name{attribute_shapefile}
\alias{attribute_shapefile}
\title{Attributing a spatial data frame using another spatial data frame}
\usage{
attribute_shapefile(spdf1, spdf2, attributefield = NULL,
  newfield = NULL)
}
\arguments{
\item{spdf1}{A SpatialPoints/PolygonsDataFrame containing the geometry to add an attribute to}

\item{spdf2}{A SpatialPoints/PolygonsDataFrame containing the geometry to add an attribute from}

\item{attributefield}{The name of the field in \code{spdf2} as a string containing the values to add to \code{spdf1}}

\item{newfield}{The name of the field in \code{spdf1} as a string to add the values from \code{spdf2$attributefield} to. If NULL, the field will use \code{attributefield}. Defaults to NULL.}
}
\value{
The original SPDF spdf1 with the new field containing the values inherited from spdf2.
}
\description{
This function will take a SpatialPoints/PolygonsDataFrame and add one attribute fields from a second SpatialPoints/PolygonsDataFrame
}
\examples{
attribute.shapefile()
}
