% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_around.R
\name{combine_designs}
\alias{combine_designs}
\title{Select points that most closely approximate the distribution of another set of points and substitute them}
\usage{
combine_designs(
  sub_points,
  template_points,
  sub_idvar = "plotid",
  template_idvar = "plotid",
  replacement_count = NULL,
  sub_counts = NULL,
  strata_polygons = NULL,
  polygons_stratavar = "stratum",
  sub_stratavar = "stratum",
  template_stratavar = "stratum",
  projection = NULL,
  iteration_limit = NULL
)
}
\arguments{
\item{sub_points}{Spatial Points Data Frame. The points you would like to substitute from by comparing against \code{template_points}.}

\item{template_points}{Spatial Points Data Frame. The points you would like to compare against \code{sub_points} in order to select a subset of those that most closely resemble the distribution of the template points.}

\item{sub_idvar}{Character string. The name of the variable in \code{sub_points@data} that contains the unique identifiers. Defaults to \code{"plotid"}.}

\item{template_idvar}{Character string. The name of the variable in \code{template_points@data} that contains the unique identifiers. Defaults to \code{"plotid"}.}

\item{replacement_count}{Optional integer. The number of template points to replace with the closest approximation from the existing points. At least one point in each stratum will be replaced and the rest of the remain replacements will be made proportionally to the weight of the strata (e.g. a stratum that makes up 1/2 of the area or contains 1/2 of the template points will also contain 1/2 of the replacements). If \code{NULL} then points will be allocated according to the vector \code{sub_counts}. Defaults to \code{NULL}.}

\item{sub_counts}{Optional named numeric vector. A vector of the number of template points to replace with the closest approximation from the existing points in each stratum. Only use if there is more than one stratum. The names of the vector must exactly match the strata of the points. If \code{NULL} then points will be allocated proportionally by area using the value in \code{replacement_count}. Defaults to \code{NULL}.}

\item{strata_polygons}{Optional Spatial Polygons Data Frame. Polygons assigned a variable with a name \code{polygons_stratavar} that contains the membership information (e.g. strata) to assign to \code{sub_points} and \code{template_points}. If \code{NULL} then no assignment will be attempted. Defaults to \code{NULL}.}

\item{polygons_stratavar}{Optional character string. If \code{strata_polygons} is not \code{NULL}, the name of the variable in \code{strata_polygons@data} that contains the membership information. Otherwise, the name of the variable in both \code{template_points@data} and \code{sub_points@data} that contains the membership information. If \code{NULL} then the points will be considered to belong to a single group. Defaults to \code{"stratum"}.}

\item{sub_stratavar}{Optional character string. The name of the variable in \code{sub_points@data} that contains the membership information. If both \code{sub_stratavar} and \code{template_stratavar} are \code{NULL} then all the points in both will be considered to belong to a single group. Defaults to \code{"stratum"}.}

\item{template_stratavar}{Optional character string. The name of the variable in \code{template_points@data} that contains the membership information. If both \code{sub_stratavar} and \code{template_stratavar} are \code{NULL} then all the points in both will be considered to belong to a single group. Defaults to \code{"stratum"}.}

\item{projection}{Optional CRS object. The projection to force all spatial objects into, e.g. \code{sp::CRS("+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0")}. If \code{NULL} then the projection from \code{template_points@proj4string} will be used. Defaults to \code{NULL}.}
}
\value{
A spatial points data frame of the \code{template_points} with points from \code{sub_points} substituted in. It will be in the projection specified by \code{projection}.
}
\description{
When creating a design that incorporates points from another design to revisit, it can be important to approximate the same spatial distribution, especially if the design is spatially balanced. This will substitute in revisit points that will cause the least shift in the spatial distribution.
}
