% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{compare_designs}
\alias{compare_designs}
\title{Comparing two design objects}
\usage{
compare_designs(design1, design2)
}
\arguments{
\item{design1}{List (see: \code{spsurvey::grts()}) or point allocation data frame (see: \code{sample.design::read_panels()}) describing a design expected to be identical to \code{design2}.}

\item{design2}{List (see: \code{spsurvey::grts()}) or point allocation data frame (see: \code{sample.design::read_panels()}) describing a design expected to be identical to \code{design1}.}
}
\value{
Data frame with three variables: panel, stratum, and error
}
\description{
Takes two different design objects formatted for \code{spsurvy::grts()} and returns a data frame detailing the differences between them as errors
}
