% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.r
\name{design_dataframe}
\alias{design_dataframe}
\title{Convert a GRTS design list to a data frame}
\usage{
design_dataframe(design)
}
\arguments{
\item{design}{A list in the format taken by \code{spsurvey::grts()}.}
}
\description{
Convert a GRTS design list to a data frame
}
