% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.search.r
\name{df.search}
\alias{df.search}
\title{Search a variable in an SPDF for values}
\usage{
df.search(df, field = "", ignore.case.field = T, regex = "",
  ignore.case.regex = T)
}
\arguments{
\item{df}{The data frame or SPDF to search through.}

\item{field}{A character string specifying the field name to search within.}

\item{ignore.case.field}{Logical. If \code{T} then search for \code{field} in the variable names in \code{spdf} is case insensitive. Defaults to \code{T}.}

\item{regex}{A character string regular expression to be passed to \code{grepl()}.}

\item{ignore.case.regex}{Logical. If \code{T} then search for \code{regex} in \code{field} is case insensitive. Defaults to \code{T}.}
}
\value{
A vector of unique values found in \code{field} matching \code{regex}.
}
\description{
A wrapper for \code{[grepl(),]} to help quickly find values matching a regular expression in a given variable in an SPDF.
}
