% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{dissolve_polygons}
\alias{dissolve_polygons}
\title{Dissolve polygons in an SPDF by an identity variable}
\usage{
dissolve_polygons(polygons, dissolve_field)
}
\arguments{
\item{polygons}{Spatial Polygons Data Frame. The polygons to dissolve. Must have an identity variable with a name patching \code{dissolve_field}.}

\item{dissolve_field}{Character string. The name of the variable in \code{polygons@data} containing the polygon identities}
}
\value{
A spatial polygons data frame with one observation per unique identity in the original \code{polygons}. The associated data frame contains only the identities in a variable named using \code{dissolve_field}.
}
\description{
Dissolve polygons together on an identity variable to produce an SPDF with one observation per identity.
}
