% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{dist_matrix}
\alias{dist_matrix}
\title{Calculate a distance matrix for a data frame of coordinates}
\usage{
dist_matrix(dataframe, x_var, y_var)
}
\arguments{
\item{dataframe}{A data frame. Must contain numeric variables corresponding to \code{x_var} and \code{y_var} containing the X and Y coordinate values.}

\item{x_var}{Character string. The name of the variable in \code{dataframe} containing the X components of the coordinates.}

\item{y_var}{Character string. The name of the variable in \code{dataframe} containing the Y components of the coordinates.}
}
\value{
A matrix of the distances between the points.
}
\description{
Given a data frame containing X and Y Cartesian coordinates, calculate a distance matrix between the points.
}
