% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{extract_poly_area}
\alias{extract_poly_area}
\title{Get the areas of polygons}
\usage{
extract_poly_area(polygons, cum_freq = TRUE)
}
\arguments{
\item{polygons}{Spatial Polygons or Spatial Polygons Data Frame. The polygons to be summarized.}

\item{cum_freq}{Logical. If \code{TRUE} then the output data frame will include the variable \code{"cum_freq"} with the cumulative frequencies of the polygon identities. Note that it sorts the polygons from smallest to largest area before doing this. Defaults to \code{TRUE}.}
}
\value{
A data frame containing the polygon index within \code{polygons@polygons[[1]]@Polygons}, and if asked for, the proportional area (\code{"area_prop"}) and the cumulative frequency (\code{"cum_freq"}).
}
\description{
Given a Spatial Polygons Data Frame, produce a data frame summarizing the areas of polygons sharing an identity
}
