% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.name.r
\name{find.name}
\alias{find.name}
\title{Find names in an object matching a pattern}
\usage{
find.name(obj, name = "", ignore.case.name = F, bookend, multiple = F)
}
\arguments{
\item{obj}{The named object (a data frame, Spatial Data Frame, list, etc.) to search for matching names.}

\item{name}{A string to search the names of \code{obj} using. This will be used as a regular expression in \code{grepl()}.}

\item{bookend}{Logical. If \code{T} then \code{^} and \code{$} will be added to the ends of the regular expression. Defaults to \code{T}.}

\item{multiple}{Logical. If \code{T} then the function can return a vector of multiple names instead of just one. If \code{F} then it will produce an error if multiple matches are found. Defaults to \code{F}.}

\item{ignore.case}{Logical. Passed to \code{grepl()} to decide if the regular expression should be case insensitive. Defaults to \code{F}.}
}
\value{
A vector of character strings.
}
\description{
Find names in an object matching a pattern
}
