% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.name.r
\name{find_name}
\alias{find_name}
\title{Find names in an object matching a pattern}
\usage{
find_name(obj, name = "", ignore_case_name = FALSE, bookend,
  multiple = FALSE)
}
\arguments{
\item{obj}{The named object (a data frame, Spatial Data Frame, list, etc.) to search for matching names.}

\item{name}{A string to search the names of \code{obj} using. This will be used as a regular expression in \code{grepl()}.}

\item{ignore_case_name}{Logical. Passed to \code{grepl()} to decide if the regular expression should be case insensitive. Defaults to \code{FALSE}.}

\item{bookend}{Logical. If \code{TRUE} then \code{^} and \code{$} will be added to the ends of the regular expression. Defaults to \code{TRUE}.}

\item{multiple}{Logical. If \code{TRUE} then the function can return a vector of multiple names instead of just one. If \code{FALSE} then it will produce an error if multiple matches are found. Defaults to \code{FALSE}.}
}
\value{
A vector of character strings.
}
\description{
Find names in an object matching a pattern
}
