% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate the geometric mean of a numeric vector}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector. The values to calculate a geometric mean from}

\item{na.rm}{Logical. If \code{TRUE} then \code{NA} values will be dropped from \code{x} before calculating the geometric mean. Defaults to \code{TRUE}}
}
\value{
The geometric mean as a single numeric value.
}
\description{
Calculate the geometric mean of a numeric vector
}
