% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grts.aim.r
\name{grts.aim}
\alias{grts.aim}
\title{Draw spatially balanced points with \code{spsurvey::grts()}, output in AIM format}
\usage{
grts.aim(design.object, design.name = "Design name",
  src.frame = "sp.object", sp.object = NULL, in.shape = NULL,
  stratum.field = "STRATUM")
}
\arguments{
\item{design.object}{A list of lists structured and named according to the requirements of \code{spsurvey::grts()}. Can be automatically constructed with \code{allocate.panels()}.}

\item{design.name}{A character string of the name to use for the design, often the name of the project.}

\item{sp.object}{If using an SPDF instead of a shapefile, an SPDF with a variable name matching \code{stratum.field} which contains values matching the strata names in \code{design.object}. Defaults to \code{NULL}.}

\item{in.shape}{If using a shapefile instead of an SPDF, a string representing the filepath to a shapefile (without file extension) to pass to \code{spsurvey::grts()} as its argument \code{in.shape}. Defaults to \code{NULL}.}

\item{stratum.field}{A character string representing the name of the variable in either \code{sp.object} or \code{in.shape} containing the strata names matching those in \code{design.object}. Defaults to \code{"STRATUM"}.}
}
\value{
A Spatial Points Data Frame of the sampling locations with the fields \code{PLOTID}, \code{STRATUM}, \code{PANEL}, \code{IntPtWt} (initial point weight), \code{xcoord}, and \code{ycoord}
}
\description{
A wrapper for \code{spsurvey::grts()} that modifies the output SPDF to match the format expected by other parts of the AIM workflow, including distribution to field crews and populating design databases.
}
