% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_around.R
\name{ranked_sort}
\alias{ranked_sort}
\title{Find optimal pairings from two sets based on their rankings}
\usage{
ranked_sort(
  match_to,
  match_from,
  match_to_idvar = "match_to_id",
  match_from_idvar = "match_from_id",
  match_to_rankvar = "match_to_rank",
  match_from_rankvar = "match_from_rank",
  iteration_limit = NULL
)
}
\arguments{
\item{match_to}{Data frame. The preferences of the members of the set to match to for the members of the set to match from. There must be one observation for each unique pairing between the two sets. The variable \code{match_to_idvar} must contain the identities of the members of the set to match to; \code{match_from_idvar}, the set to match from; and \code{match_to_rankvar} the rank preference of the MATCH FROM member for the MATCH FROM member.}

\item{match_from}{Data frame. The preferences of the members of the set to match from for the members of the set to match to. There must be one observation for each unique pairing between the two sets. The variable \code{match_to_idvar} must contain the identities of the members of the set to match to; \code{match_from_idvar}, the set to match from; and \code{match_from_rankvar} the rank preference of the MATCH FROM member for the MATCH TO member.}

\item{match_to_idvar}{Character string. The name of the variable in both \code{match_to} and \code{match_from} that contains the identities of the members in the set to match to. Defaults to \code{"match_to_id"}.}

\item{match_from_idvar}{Character string. The name of the variable in both \code{match_to} and \code{match_from} that contains the identities of the members in the set to match from. Defaults to \code{"match_from_id"}.}

\item{match_to_rankvar}{Character string. The name of the variable in \code{match_to} that contains the rank preference of the match to members for the match from members as a relative numeric value (lower values are more preferred). Defaults to \code{"match_to_rank"}.}

\item{match_from_rankvar}{Character string. The name of the variable in \code{match_from} that contains the rank preference of the match from members for the match to members as a relative numeric value (lower values are more preferred). Defaults to \code{"match_from_rank"}.}
}
\value{
A data frame with the variables \code{match_to_id} and \code{match_from_id} containing the paired members of the two sets.
}
\description{
For two sets of unique identifiers, given each member's ranked preference for every member of the other set, find optimal pairings. The two sets are referred to as \code{match_to} and \code{match_from}. All the members of \code{match_to} will be paired to members of \code{match_from}, but if there are more members of \code{match_from} than \code{match_to} there will be unpaired, discarded members of \code{match_from}. The inputs are two data frames, one where each observation is a unique combination of each member of \code{match_to} and \code{match_from} and the preference of that \code{match_to} member for \code{match_from}, numeric with lower values treated as more preferred. The other data frame contains the preferences of the \code{match_from} members for \code{match_to} members.
}
