% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.raster.R
\name{raster_spdf}
\alias{raster_spdf}
\title{Convert a raster to Spatial Polygons restriced to a frame}
\usage{
raster_spdf(raster, frame_spdf, buffer.meters = 100)
}
\arguments{
\item{raster}{Raster. The raster to be clipped}

\item{frame_spdf}{Spatial polygons or spatial polygons data frame. The frame to clip the raster to.}

\item{buffer.meters}{Numeric. Number of meters to buffer the frame by before masking the raster. Make sure this is slightly larger than the raster cell size to make sure that no cells are lost before vectorization and clipping to the actual frame.}
}
\description{
Convert a raster to Spatial Polygons restriced to a frame
}
