% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.r
\name{read_panels}
\alias{read_panels}
\title{Use a data frame of strata and point counts to build a design object for \code{spsurvey::grts()}}
\usage{
read_panels(dataframe, stratum_field = "stratum", panel_names = NULL,
  oversample_field = NULL, oversample_proportion = 0.25,
  oversample_min = 3)
}
\arguments{
\item{dataframe}{Data frame. This must have at least a variable for the strata identities and one variable for each panel, e.g. "Stratum", "Year1", "Year2", "Year3", where each row is a stratum and the number of points desired in each panel for that stratum.}

\item{stratum_field}{Character string. This must exactly match the name of the variable in \code{dataframe} that contains the stratum identities. Defaults to \code{"stratum"}.}

\item{panel_names}{Optional vector of character strings. Necessary if the data frame given in \code{dataframe} has additional fields beyond the ones for the strata, panels, and oversample. This character vector specifies which variables correspond to panels and must match those names exactly.}

\item{oversample_field}{Optional character string. If used, this must exactly match the name of the variable in \code{dataframe} that contains the TOTAL number of oversample points desired for the stratum in the design. If not provided or specified as \code{NULL}, oversample point counts will be calculated using \code{oversample_proportion} and \code{oversample_min}. Defaults to \code{NULL}.}

\item{oversample_proportion}{Optional numeric value. If not providing an oversample point count in a variable specified by \code{oversample_field}, this must be between 0 and 1, representing the minimum relative proportion of oversample points to allocate per stratum per panel using the formula \code{panel point count * min.oversample_proportion}. Defaults to \code{0.25}.}

\item{oversample_min}{Optional numeric value. If not providing an oversample point count in a variable specified by \code{oversample_field}, this must be a positive integer, representing the minimum number of oversample points to allocate per stratum per panel. This is only used if it is greater than \code{ppanel point count * min.oversample_proportion}. Defaults to \code{3}.}
}
\description{
Use a data frame of strata and point counts to build a design object for \code{spsurvey::grts()}
}
