% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{select_from_distribution}
\alias{select_from_distribution}
\title{Select polygon from a probability distribution}
\usage{
select_from_distribution(
  dataframe,
  value,
  prob_var = "cum_freq",
  id_var = "id"
)
}
\arguments{
\item{dataframe}{A data frame. Must contain an identity variable with a name matching \code{id_var} and a probability variable with a name matching \code{prob_var}. Will be sorted by ascending probability.}

\item{value}{Numeric. A single numeric value to compare against the probabilities in \code{dataframe}.}

\item{prob_var}{Character string. Must match the name of the variable in \code{dataframe} that contains the probability values. Defaults to \code{"cum_freq"}.}

\item{id_var}{Character string. Must match the name of the variable in \code{dataframe} that contains the identities. Defaults to \code{"id"}.}
}
\value{
The identity value from \code{dataframe} with the smallest probabilty value greater than \code{value} OR the first identity value if none were greater than \code{value}.
}
\description{
Given a data frame of IDs with associated probabilities and a number, select the ID with the smallest probability value greater than than the number given.
}
