% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring_resources_functions.R
\name{strat_panel.fcn}
\alias{strat_panel.fcn}
\title{Strata time}
\usage{
strat_panel.fcn(master_df, target_panel_size, wgt_var = "wgt_km",
  wgt_units = "km", wgt_type = "Length", strata_var = NULL,
  id_var = "site.id", x_var = "x.coord", y_var = "y.coord",
  rerandomize = TRUE, prob_var = NULL, legacy_df = NULL,
  legacy_first = FALSE, oversample_pct = NULL)
}
\arguments{
\item{master_df}{master sample data frame}

\item{target_panel_size}{Target panel size. A named vector of point counts where the names correspond to the strata found in \\code{master_df[[strata_var]]}.}

\item{wgt_var}{Name of weight column}

\item{wgt_units}{Units of weight column}

\item{wgt_type}{Specific type of weight column (lenght, area, etc.)}

\item{strata_var}{Name of stratification column}

\item{rerandomize}{Re-randomize points? (TRUE/FALSE)}

\item{prob_var}{Name of column of probabilities}

\item{legacy_df}{The legacy data frame object (otherwise NULL)}

\item{legacy_first}{Should legacy points be selected first? (TRUE/FALSE)}

\item{oversample_pct}{The percent of points that should be oversampled}
}
\value{
this function provides a list of messages (?) and the resulting points for all stratum and panel draws
}
\description{
Do something with strata, I guess????
}
