% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{get_coords}
\alias{get_coords}
\title{Add coordinates to the data slot of a SpatialPointsDataFrame}
\usage{
get_coords(
  points,
  x_var = "XMETERS",
  y_var = "YMETERS",
  projection = CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")

)
}
\arguments{
\item{points}{A SpatialPointsDataFrame.}

\item{x_var}{Character string. The name of the variable to add the x component of the coordinates to in \code{points@data}. If the variable already exists, it will be overwritten. Defaults to \code{"XMETERS"}.}

\item{y_var}{Character string. The name of the variable to add the y component of the coordinates to in \code{points@data}. If the variable already exists, it will be overwritten. Defaults to \code{"YMETERS"}.}

\item{projection}{CRS object. The projection to use when determining the coordinates. Defaults to \code{sp::CRS("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")}, Albers equal area.}
}
\value{
The SPDF provided as \code{points} with the coordinates added to \code{@data$x_var} and \code{@data$y_var}.
}
\description{
Add coordinates to the data slot of a SpatialPointsDataFrame
}
