% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grts.aim.r
\name{grts_aim}
\alias{grts_aim}
\title{Draw spatially balanced points with \code{spsurvey::grts()}, output in AIM format}
\usage{
grts_aim(
  design_object = NULL,
  n_base = NULL,
  n_over = NULL,
  project_name = "Design name",
  frame,
  stratum_var = "STRATUM",
  seed_number = NULL,
  projection = NULL
)
}
\arguments{
\item{design_object}{Optional if not providing \code{n_base} and \code{n_over}. A list of lists structured and named according to the requirements of \code{spsurvey::grts()} prior to version 5.3.0. Can be automatically constructed with \code{allocate.panels()}.}

\item{n_base}{Optional if not providing \code{design_object}. A named numeric value for the numbers of base points to draw in each stratum in the design. Names must match the values in \code{frame} in the variable specified by \code{stratum_field}.}

\item{n_over}{Optional if not providing \code{design_object}. A named numeric value for the numbers of oversample (spare) points to draw  in each stratum in the design. Names must match the values in \code{frame} in the variable specified by \code{stratum_field}.}

\item{project_name}{A character string of the name to use for the design, often the name of the project.}

\item{frame}{Polygon sf object with a variable name matching \code{stratum_field} which contains values matching the strata names in \code{design_object} OR in \code{n_base} and \code{n_over}.}

\item{stratum_var}{A character string representing the name of the variable in \code{frame} containing the strata names matching those in \code{design_object} OR in \code{n_base} and \code{n_over}. Defaults to \code{"STRATUM"}.}

\item{seed_number}{An optional numeric value to be passed to \code{set.seed()} to make the output reproducible. If \code{NULL} then a random seed will be used. Defaults to \code{NULL}.}

\item{projection}{Optional character string or CRS object. The projection to reproject the frame into if no \code{NULL}. Defaults to \code{NULL}.}
}
\value{
A Spatial Points Data Frame of the sampling locations with the fields \code{PLOTID}, \code{STRATUM}, \code{PANEL}, \code{IntPtWt} (initial point weight), \code{xcoord}, and \code{ycoord}
}
\description{
A wrapper for \code{spsurvey::grts()} that modifies the output SPDF to match the format expected by other parts of the AIM workflow, including distribution to field crews and populating design databases.
}
