% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{search}
\alias{search}
\title{Find values in a data frame column/variable that match a list or vector}
\usage{
search(dataframe = NULL, values, namestring = "", use_grepl = FALSE,
  ignore_case_namestring = FALSE, ignore_case_values = FALSE,
  bookend_namestring = FALSE, bookend_values = FALSE)
}
\description{
Return all unique values from a column/variable in a data frame which match values in a list or vector.
}
