% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{test_points}
\alias{test_points}
\title{Test points spatial balance against sets of random points}
\usage{
test_points(number = 500, pts_spdf, aoi_spdf, type = 1,
  seed_number = 420)
}
\arguments{
\item{number}{Numeric. The number of sets to generate to compare. Defaults to \code{3}.}

\item{pts_spdf}{Spatial Points Data Frame. The points that are being compared against.}

\item{aoi_spdf}{Spatial Polygons Data Frame. Polygons describing the boundaries of the area of interest that corresponds to \code{pts_spdf}.}

\item{type}{Numeric. Use \code{1} when aoi_spdf is a dissolved set of polygons and \code{2} when aoi_spdf is an undissolved set of polylines. Defaults to \code{1}}

\item{seed_number}{Numeric. The number to use in \code{set.seed()} for reproducibility. Defaults to \code{1}.}
}
\value{
Named numeric vector. The value for \code{"p_arith"} is the proportion of comparisons that had a higher arithmetic mean nearest neighbor distance than \code{pts_spdf} and \code{"p_geom"} is the proportion of comparisons that had a higher geometric mean nearest neighbor distance.
}
\description{
Compare a set of points to sets of random points generated from the same polygon geometry and report back the proportion of random sets which had higher mean distance to nearest neighboring point. Assuming that a higher mean distance to nearest neighbor indicates greater spatial balance, this proportion can be treated as a "probabilty that the points in \code{spdf} are not spatially balanced." Whatever is provided as \code{spdf} will be dissolved without regard for the data slot, so if you want to test subsets of the polygons, each test will need to be a separate function call provided only the relevant subset as \code{spdf}, e.g. in the case of wanting to test individual strata stored as a single SPDF you would need to call this function for each stratum (probably in a \code{lapply()} or a loop).
}
